+++ packages/expo-build-properties/src/android.ts
@@ -100,6 +100,18 @@ export const withAndroidBuildProperties = createBuildGradlePropsConfigPlugin<Plu
       propValueGetter: (config) => config.android?.enableBundleCompression?.toString(),
     },
     {
+      propName: 'expo.gif.enabled',
+      propValueGetter: (config) => config.android?.gifEnabled?.toString(),
+    },
+    {
+      propName: 'expo.webp.enabled',
+      propValueGetter: (config) => config.android?.webpEnabled?.toString(),
+    },
+    {
+      propName: 'expo.webp.animated',
+      propValueGetter: (config) => config.android?.webpAnimated?.toString(),
+    },
+    {
       propName: 'reactNativeArchitectures',
       propValueGetter: (config) => config.android?.buildArchs?.join(','),
     },